# internal/api/openapi/v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cannon Content API",
    "version": "1.0.0",
    "description": "Headless read-only content and account self-service API for Cannon CMS. All endpoints require the X-Cannon-API-Key header. User-specific visibility and account routes also accept Authorization: Bearer <JWT>."
  },
  "servers": [{ "url": "/api/v1" }],
...
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Cannon-API-Key",
        "description": "Application credential issued in Admin → API → Credentials"
      },
      "UserBearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "User access token from POST /auth/login"
      }
    },
    "parameters": {
      "Page": {
        "name": "page",
        "in": "query",
        "description": "Page number (1-based).",
        "schema": { "type": "integer", "minimum": 1, "default": 1 }
      },
      "PageSize": {
        "name": "page_size",
        "in": "query",
        "description": "Items per page (max 100).",
        "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 }
      },
      "SearchQuery": {
        "name": "q",
        "in": "query",
        "description": "Search keywords.",
        "schema": { "type": "string" }
      },
      "Locale": {
        "name": "locale",
        "in": "query",
        "description": "Content locale when multilingual mode is enabled.",
        "schema": { "type": "string", "example": "en-US" }
      },
      "CategoryIDQuery": {
        "name": "category_id",
        "in": "query",
        "description": "Filter items assigned to this category ID.",
        "schema": { "type": "integer", "minimum": 1 }
      },
      "CategorySlugQuery": {
        "name": "category_slug",
        "in": "query",
        "description": "Filter items in this category slug and its descendants.",
        "schema": { "type": "string", "example": "news" }
      },
      "IncludeDescendants": {
        "name": "include_descendants",
        "in": "query",
        "description": "When used with category_id, include items in child categories.",
        "schema": { "type": "string", "enum": ["1"] }
      },
      "TagIDQuery": {
        "name": "tag_id",
        "in": "query",
        "description": "Filter items with this tag ID.",
        "schema": { "type": "integer", "minimum": 1 }
      },
      "Featured": {
        "name": "featured",
        "in": "query",
        "description": "Set to 1 to return featured items only.",
        "schema": { "type": "string", "enum": ["1"] }
      },
...
      "CategoryIDPath": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Category ID.",
        "schema": { "type": "integer", "minimum": 1 }
      },
      "CategorySlugPath": {
        "name": "slug",
        "in": "path",
        "required": true,
        "description": "Category slug path (may include slashes, e.g. news/local).",
        "schema": { "type": "string", "example": "news" }
      },
      "ItemIDPath": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Item ID.",
        "schema": { "type": "integer", "minimum": 1 }
      },
      "ItemSlugPath": {
        "name": "slug",
        "in": "path",
        "required": true,
        "description": "Item slug for the active locale.",
        "schema": { "type": "string" }
      },
...
    }
  },
  "security": [{ "ApiKeyAuth": [] }],
  "paths": {
...
    "/items": {
      "get": {
        "operationId": "listItems",
        "tags": ["Items"],
        "summary": "List visible published items",
        "parameters": [
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/PageSize" },
          { "$ref": "#/components/parameters/SearchQuery" },
          { "$ref": "#/components/parameters/Locale" },
          { "$ref": "#/components/parameters/CategoryIDQuery" },
          { "$ref": "#/components/parameters/CategorySlugQuery" },
          { "$ref": "#/components/parameters/IncludeDescendants" },
          { "$ref": "#/components/parameters/TagIDQuery" },
          { "$ref": "#/components/parameters/Featured" }
        ],
        "responses": { "200": { "description": "Paginated items" } }
      }
    },
    "/items/{id}": {
      "get": {
        "operationId": "getItemByID",
        "tags": ["Items"],
        "summary": "Item by ID",
        "parameters": [{ "$ref": "#/components/parameters/ItemIDPath" }],
        "responses": { "200": { "description": "Item detail" }, "404": { "description": "Not found or not visible" } }
      }
    },
    "/items/by-slug/{slug}": {
      "get": {
        "operationId": "getItemBySlug",
        "tags": ["Items"],
        "summary": "Item by slug",
        "parameters": [{ "$ref": "#/components/parameters/ItemSlugPath" }],
        "responses": { "200": { "description": "Item detail" }, "404": { "description": "Not found or not visible" } }
      }
    },
...
    "/categories/{id}": {
      "get": {
        "operationId": "getCategoryByID",
        "tags": ["Categories"],
        "summary": "Category by ID",
        "parameters": [{ "$ref": "#/components/parameters/CategoryIDPath" }],
        "responses": { "200": { "description": "Category detail with direct children" }, "404": { "description": "Not found or not visible" } }
      }
    },
    "/categories/by-slug/{slug}": {
      "get": {
        "operationId": "getCategoryBySlug",
        "tags": ["Categories"],
        "summary": "Category by slug",
        "parameters": [{ "$ref": "#/components/parameters/CategorySlugPath" }],
        "responses": { "200": { "description": "Category detail with direct children" }, "404": { "description": "Not found or not visible" } }
      }
    },
    "/categories/{id}/items": {
      "get": {
        "operationId": "listCategoryItems",
        "tags": ["Categories"],
        "summary": "Items in a category (includes descendants)",
        "parameters": [
          { "$ref": "#/components/parameters/CategoryIDPath" },
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/PageSize" },
          { "$ref": "#/components/parameters/SearchQuery" }
        ],
        "responses": { "200": { "description": "Paginated items" }, "404": { "description": "Category not found or not visible" } }
      }
    },
...
  }
}